theme switch and user dropdown to component
